immutable_chunkmap
Module set

Structs

Set

This set uses a similar strategy to BTreeSet to ensure cache efficient performance on modern hardware while still providing log(N) get, insert, and remove operations.

SetIter

WeakSetRef

Constants

DEFAULT_SIZE

Type Aliases

SetL

set with a larger chunk size, faster to search, slower to update

SetM

set with the default chunk size, a good balance of search and update performance

SetS

set with a smaller chunk size, faster to update, slower to search

WeakSetRefL

WeakSetRefM

WeakSetRefS