immutable_chunkmap

Module map

Structs

Map

This Map uses a similar strategy to BTreeMap to ensure cache efficient performance on modern hardware while still providing log(N) get, insert, and remove operations.

WeakMapRef

A weak reference to a map.

Constants

DEFAULT_SIZE

Type Aliases

MapL

Map using a larger chunk size, faster to search, slower to update

MapM

Map using the default chunk size, a good balance of update and search

MapS

Map using a smaller chunk size, faster to update, slower to search

WeakMapRefL

WeakMapRefM

WeakMapRefS